+++ src/pages/Services/Store/components/DetailForm.tsx
@@ -3,7 +3,7 @@ import CustomForm from 'components/common/CustomForm';
 import { Divider } from 'antd';
 import { FormInstance } from 'antd/es/form/Form';
 import STORE_OPTION from 'constant/store';
-import CustomProList from 'components/common/CustomProForm/CustomProList';
+import CustomProList from 'components/menu/CustomProForm/CustomProList';
 import * as S from '../StoreDetail.style';
 import StoreCategory from './StoreCategory';
 import OpenTimeForm from './OpenTimeForm';
